fix for spamming moves crash (parsing error on client side)
This bug was occuring because we were clearing the entire buffer instead
of only the part we read (until the \n).

Because of that, if an async_read_until was started and read more than
up to the \n, the completion handler was then called, and we were
clearing the entire buffer, including what was after the \n and that has
already been read, making those bytes disappear.
We were then reading again up to the next \n, but since we had lost the
firt part of the packet, the json was incomplete.