removed "GUI" in the names where it was useless
We already know it's GUI with the namespace, therefore writing GUI in
the name of each type is useless.
Adding it in the member variable's names is useless too since there
isn't any GUI component that would have a TUI component member, so we
already know it's GUI for sure.